feat(auth): add OTPless API wrapper for OTP initiation and verification
- Implemented `otpless.js` to handle OTP flows for customers and delivery partners.
- Added functions to initiate and verify OTPs, with support for SMS and WhatsApp channels.
- Introduced development mode for testing without sending actual OTPs.

feat(token): create access and refresh token helpers

- Created `token.js` to manage JWT access and refresh tokens for delivery-partner authentication.
- Implemented signing and verification functions for both token types.

chore(deps): update dependencies in yarn.lock

- Added new dependencies for axios, form-data, and others to support the new authentication flows.
- Updated existing dependencies to their latest versions for improved stability and security.

+++ models/store_orders_details.js
@@ -15,7 +15,8 @@ const OrderDetails = sequelize.define(
       allowNull: false
     },
     product_id: {
-      type: DataTypes.SMALLINT,
+      // Must match store_products.product_id (INTEGER) for the FK to build.
+      type: DataTypes.INTEGER,
       allowNull: false
     },
     product_qty: {
